for fixes for thread handling
